🚚(backend) move react_locale template tag to core app
The react_locale template tag is useful any time a React-powered
component needs to be included in a page. This is not limited to
the Search application any more.

Therefore, it makes more sense to move it to the core app so it
is available throughout richie. We also renamed the file from
`search_tags` to `react_locale` so the way to enable it in a template
is more self-evident.
