+++ src/richie/apps/core/templatetags/react_locale.py
