+++ src/richie/apps/search/templates/search/search.html
@@ -1,5 +1,5 @@
 {% extends "richie/base.html" %}
-{% load cms_tags i18n search_tags static %}
+{% load cms_tags i18n react_locale static %}
 {% get_current_language as LANGUAGE_CODE %}
 
 {% block meta %}
